Controller consolidation (#297, #299, #261) (#91)
* feat: consolidate Google integration into GoogleController (#297)

Move Google-related actions from AdminController, AdminEmailController,
BoardController, TeamController, and HumanController into a new
GoogleController at /Google. Team-level resource linking remains in
TeamAdminController. Update all views and nav links to point to new routes.

* fix: remove AdminEmailController and old view files (#297)

Delete AdminEmailController (fully replaced by GoogleController.Accounts*)
and remove old view files that were copied to Views/Google/.

* feat: consolidate email operations into EmailController (#299)

Move EmailOutbox, Pause/Resume, Retry/Discard, and EmailPreview from
AdminController to new EmailController at /Email. Facilitated messaging
(HumanController.SendMessage, CampController.Contact) remains contextual
by design. AdminController is now significantly smaller.

* refactor: rename HumanController actions for clarity (#261)

Rename View -> HumanProfile and HumanDetail -> HumanProfileAdmin to
accurately describe the pages they serve. Update all RedirectToAction,
asp-action, and Url.Action references. Routes unchanged.

* fix: remove unused dependencies from BoardController and HumanController

Clean up unused fields and constructor parameters left over from
controller consolidation.

---------
