+++ src/Humans.Web/Views/Email/EmailOutbox.cshtml
@@ -43,7 +43,7 @@
                 </button>
             </form>
         }
-        <a asp-action="Index" class="btn btn-outline-secondary">Back</a>
+        <a asp-controller="Admin" asp-action="Index" class="btn btn-outline-secondary">Back</a>
     </div>
 </div>
