+++ src/Humans.Web/Views/Google/AllGroups.cshtml
@@ -47,7 +47,7 @@
                 <i class="fa-solid fa-list-check me-1"></i>Per-Team Drift View
             </a>
             <a asp-action="Index" class="btn btn-outline-secondary btn-sm">
-                <i class="fa-solid fa-arrow-left me-1"></i>Back to Admin
+                <i class="fa-solid fa-arrow-left me-1"></i>Back to Google
             </a>
         </div>
     </div>
