+++ src/Humans.Web/Views/Google/GroupSettingsResults.cshtml
@@ -11,7 +11,7 @@
                 <i class="fa-solid fa-layer-group me-1"></i>All Domain Groups
             </a>
             <a asp-action="Index" class="btn btn-outline-secondary btn-sm">
-                <i class="fa-solid fa-arrow-left me-1"></i>Back to Admin
+                <i class="fa-solid fa-arrow-left me-1"></i>Back to Google
             </a>
         </div>
     </div>
