+++ src/Humans.Web/Views/Google/SyncResults.cshtml
@@ -7,7 +7,7 @@
     <div class="d-flex justify-content-between align-items-center mt-4 mb-3">
         <h1>Sync Results</h1>
         <a asp-action="Index" class="btn btn-outline-secondary btn-sm">
-            <i class="fa-solid fa-arrow-left me-1"></i>Back to Admin
+            <i class="fa-solid fa-arrow-left me-1"></i>Back to Google
         </a>
     </div>
