+++ src/Humans.Web/Views/Human/HumanProfileAdmin.cshtml
@@ -11,7 +11,7 @@
 <nav aria-label="breadcrumb">
     <ol class="breadcrumb">
         <li class="breadcrumb-item"><a asp-action="Humans">Humans</a></li>
-        <li class="breadcrumb-item"><a asp-action="View" asp-route-id="@Model.UserId">@Model.DisplayName</a></li>
+        <li class="breadcrumb-item"><a asp-action="HumanProfile" asp-route-id="@Model.UserId">@Model.DisplayName</a></li>
         <li class="breadcrumb-item active" aria-current="page">Admin</li>
     </ol>
 </nav>
@@ -308,7 +308,7 @@
         <div class="card mb-4">
             <div class="card-header">Google Sync</div>
             <div class="card-body">
-                <a asp-action="HumanGoogleSyncAudit" asp-route-id="@Model.UserId"
+                <a asp-controller="Google" asp-action="HumanGoogleSyncAudit" asp-route-id="@Model.UserId"
                    class="btn btn-outline-secondary w-100">View Google Sync Audit</a>
             </div>
         </div>
@@ -325,7 +325,7 @@
                 }
                 else
                 {
-                    <form asp-action="ProvisionEmail" asp-route-id="@Model.UserId" method="post">
+                    <form asp-controller="Google" asp-action="ProvisionEmail" asp-route-id="@Model.UserId" method="post">
                         @Html.AntiForgeryToken()
                         <div class="input-group mb-2">
                             <input type="text" name="emailPrefix" class="form-control form-control-sm"
