+++ src/Humans.Web/Views/Human/Humans.cshtml
@@ -149,7 +149,7 @@
                                 }
                             </td>
                             <td class="text-end">
-                                <a asp-action="HumanDetail" asp-route-id="@member.Id" class="btn btn-sm btn-outline-primary">
+                                <a asp-action="HumanProfileAdmin" asp-route-id="@member.Id" class="btn btn-sm btn-outline-primary">
                                     @Localizer["MyTeams_View"]
                                 </a>
                             </td>
