+++ src/Humans.Web/Views/Team/Index.cshtml
@@ -18,7 +18,7 @@
             @if (canViewSync)
             {
                 <a asp-action="Summary" class="btn btn-outline-secondary">Summary</a>
-                <a asp-action="Sync" class="btn btn-outline-info">Sync Status</a>
+                <a asp-controller="Google" asp-action="Sync" class="btn btn-outline-info">Sync Status</a>
             }
             @if (Model.CanCreateTeam)
             {
